fix multiple choice tests after refactoring
